% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_harmonise.R
\name{allele_harmonise}
\alias{allele_harmonise}
\title{allele_harmonise: Function to harmonize allele coding between a reference panel and GWAS/eQTL summary data}
\usage{
allele_harmonise(ref_panel, gwas_data)
}
\arguments{
\item{ref_panel}{A data.frame or data.table representing an LD reference panel. It must contain columns "SNP", "A1" (effect allele), and "A2" (other allele).}

\item{gwas_data}{A data.frame or data.table representing GWAS/eQTL summary data. It must contain columns "SNP", "A1", "A2", and "Zscore".}
}
\value{
A data.table with harmonized alleles and adjusted Z-scores. The output contains the harmonized alleles from the reference panel and the original alleles from the GWAS/eQTL data as separate columns.
}
\description{
This function harmonizes the allele coding between an LD reference panel and GWAS/eQTL summary data to ensure consistency in the effect allele (A1) and the other allele (A2). The Z-scores in the GWAS/eQTL summary data are adjusted accordingly if the alleles are flipped.
}
