% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_h2.R
\name{h2_liability}
\alias{h2_liability}
\title{Convert Heritability to Liability Scale}
\usage{
h2_liability(h2, sample_prev, population_prev)
}
\arguments{
\item{h2}{(numeric) Estimate of observed-scale heritability}

\item{sample_prev}{(numeric) Proportion of cases in the current sample}

\item{population_prev}{(numeric) Population prevalence of trait}
}
\value{
A list containing liability-scale heritability, standard error, and p-value
}
\description{
\code{h2_liability()} converts heritability estimates from the observed to liability scale.
}
\examples{
h2_liability(0.28, 0.1, 0.05)

}
