% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_h2.R
\name{ldsc_h2}
\alias{ldsc_h2}
\title{Estimate heritability}
\usage{
ldsc_h2(
  munged_sumstats,
  ancestry,
  sample_prev = NA,
  population_prev = NA,
  ld,
  wld,
  n_blocks = 200,
  chisq_max = NA
)
}
\arguments{
\item{munged_sumstats}{Either a dataframe, or a path to a file containing munged summary statistics. Must contain at least columns named \code{SNP} (rsid), \code{A1} (effect allele), \code{A2} (non-effect allele), \code{N} (total sample size) and \code{Z} (Z-score)}

\item{ancestry}{(character) One of \code{AFR}, \code{AMR}, \code{CSA}, \code{EAS}, \code{EUR}, or \code{MID}, which will utilize the appropriate built-in \code{ld} and \code{wld} files from Pan-UK Biobank. If empty or \code{NULL}, the user must specify paths to \code{ld} and \code{wld} files.}

\item{sample_prev}{(numeric) For binary traits, this should be the prevalence of cases in the current sample, used for conversion from observed heritability to liability-scale heritability. The default is \code{NA}, which is appropriate for quantitative traits or estimating heritability on the observed scale.}

\item{population_prev}{(numeric) For binary traits, this should be the population prevalence of the trait, used for conversion from observed heritability to liability-scale heritability. The default is \code{NA}, which is appropriate for quantitative traits or estimating heritability on the observed scale.}

\item{ld}{(character) Path to directory containing ld score files, ending in \verb{*.l2.ldscore.gz}. Default is \code{NA}, which will utilize the built-in ld score files from Pan-UK Biobank for the ancestry specified in \code{ancestry}.}

\item{wld}{(character) Path to directory containing weight files. Default is \code{NA}, which will utilize the built-in weight files from Pan-UK Biobank for the ancestry specified in \code{ancestry}.}

\item{n_blocks}{(numeric) Number of blocks used to produce block jackknife standard errors. Default is \code{200}}

\item{chisq_max}{(numeric) Maximum value of Z^2 for SNPs to be included in LD-score regression. Default is to set \code{chisq_max} to the maximum of 80 and N*0.001.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing heritability information. If \code{sample_prev} and \code{population_prev} were provided, the heritability estimate will also be returned on the liability scale.
}
\description{
\code{ldsc_h2()} uses ldscore regression to estimate the heritability of a trait from GWAS summary statistics and reference LD information.
}
\examples{
\donttest{
ldsc_h2(sumstats_munged_example(dataframe = TRUE), ancestry = "EUR")
}
}
