% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_rg.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Estimate cross-trait genetic correlations (Robust Version)}
\usage{
ldsc_rg(
  munged_sumstats,
  ancestry,
  sample_prev = NA,
  population_prev = NA,
  ld,
  wld,
  n_blocks = 200,
  chisq_max = NA,
  chr_filter = seq(1, 22, 1)
)
}
\arguments{
\item{munged_sumstats}{(list) A named list of dataframes, or paths to files containing munged summary statistics. Each set of munged summary statistics contain at least columns named \code{SNP} (rsid), \code{A1} (effect allele), \code{A2} (non-effect allele), \code{N} (total sample size) and \code{Z} (Z-score)}

\item{ancestry}{(character) One of "AFR", "AMR", "CSA", "EAS", "EUR", or "MID", which will utilize the appropriate built-in \code{ld} and \code{wld} files from Pan-UK Biobank. If empty or \code{NULL}, the user must specify paths to \code{ld} and \code{wld} files.}

\item{sample_prev}{(list) A named list containing the prevalence of cases in the current sample, used for conversion from observed heritability to liability-scale heritability. The default is \code{NA}, which is appropriate for quantitative traits or estimating heritability on the observed scale.}

\item{population_prev}{(list) A named list containing the population prevalence of the trait, used for conversion from observed heritability to liability-scale heritability. The default is \code{NA}, which is appropriate for quantitative traits or estimating heritability on the observed scale.}

\item{ld}{(character) Path to directory containing ld score files, ending in \verb{*.l2.ldscore.gz}. Default is \code{NA}, which will utilize the built-in ld score files from Pan-UK Biobank for the ancestry specified in \code{ancestry}.}

\item{wld}{(character) Path to directory containing weight files. Default is \code{NA}, which will utilize the built-in weight files from Pan-UK Biobank for the ancestry specified in \code{ancestry}.}

\item{n_blocks}{(numeric) Number of blocks used to produce block jackknife standard errors. Default is \code{200}}

\item{chisq_max}{(numeric) Maximum value of Z^2 for SNPs to be included in LD-score regression. Default is to set \code{chisq_max} to the maximum of 80 and N*0.001.}

\item{chr_filter}{(numeric vector) Chromosomes to include in analysis. Separating even/odd chromosomes may be useful for exploratory/confirmatory factor analysis.}
}
\value{
A list of class \code{ldscr_list} containing heritablilty and genetic correlation information
\itemize{
\item \code{h2} = \link[tibble:tibble-package]{tibble} containing heritability information for each trait. If \code{sample_prev} and \code{population_prev} were provided, the heritability estimates will also be returned on the liability scale.
\item \code{rg} = \link[tibble:tibble-package]{tibble} containing pairwise genetic correlations information.
\item \code{raw} = A list of correlation/covariance matrices
}
}
\description{
\code{ldsc_rg()} uses ldscore regression to estimate the pairwise genetic correlations between traits. The function relies on named lists of traits, sample prevalences, and population prevalences. The name of each trait should be consistent across each argument.
}
\details{
This function estimates the pairwise genetic correlations between an arbitrary number of traits. The function also estimates heritability for each individual trait. There is a \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} method for visualizing a heatmap of the results.

This version handles cases where traits have non-positive heritability estimates more gracefully by returning NA values for correlations involving such traits.
}
