% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldscR.R
\name{ldscR}
\alias{ldscR}
\title{Estimate Heritability and Genetic Correlation Matrix Using LDSC estimated from sample LD matrix estimate}
\usage{
ldscR(GWAS_List, LDSC, Boundary = F, zsquare_thresh = 500, cov_thresh = 300)
}
\arguments{
\item{GWAS_List}{A list of data.frames where each data.frame contains GWAS summary statistics for a trait. Each data.frame should include columns for SNP identifiers, Z-scores of effect size estimates, sample sizes (N), effect allele (A1), and reference allele (A2).}

\item{LDSC}{A data.frame containing LD Score Regression (LDSC) estimates. It should include LDSC scores and other necessary metrics for the analysis.}

\item{Boundary}{A list of the information of upper and lower boundaries of parameter estimates. It typically includes intercept.lower (he lower boundary of the intercept estimate), intercept.upper (the upper boundary of the intercept estimate), h2.upper (the upper boundary of the heritability estimate), ecov.lower (the lower boundary of the estimation error covariance estimate), ecov.upper (the upper boundary of the estimation error covariance estimate), gcov.lower (the lower boundary of the genetic covariance estimate), gcov.upper (the upper boundary of the genetic covariance estimate).}

\item{zsquare_thresh}{Threshold on the z-score square in heritability estimation.}

\item{cov_thresh}{Threshold on the abs(z-score1 * z-score2) in genetic covariance estimation.}

\item{min.eps}{A lower boundary of the eigenvalues in genetic correlation matrix estimate and estimation error correlation matrix estimate.}
}
\value{
A list containing the following elements:
\itemize{
  \item{GCovEst}{Estimated genetic covariance matrix.}
  \item{GCovSE}{Standard errors of the estimated genetic covariance matrix.}
  \item{ECovEst}{Estimated error covariance matrix.}
  \item{ECovSE}{Standard errors of the estimated error covariance matrix.}
  \item{Estimate.ini}{Initial estimates with all weights being 1.}
  \item{Computing.time}{Computing time in each stage.}
}
}
\description{
The \code{ldscR} function estimates heritability and genetic correlation matrices using Linkage Disequilibrium Score Regression (LDSC). It processes GWAS summary statistics, harmonizes alleles with a reference panel, and computes genetic covariance and error covariance matrices.
}
\details{
The \code{ldscR} function is designed for advanced genetic statistics and requires a good understanding of GWAS summary statistics, LDSC methodology, and statistical genetics. Users should ensure that input data is correctly formatted and that they understand the implications of the estimates produced by the function.
}
\examples{
data(hapmap3)
data(EURLDSC)
ref_panel <- hapmap3
LDSC <- EURLDSC
results <- ldscR(GWAS_List,LDSC)

}
