% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_h2.R
\name{sumstats_munged_example}
\alias{sumstats_munged_example}
\title{Example munged dataframe}
\usage{
sumstats_munged_example(example, dataframe = TRUE)
}
\arguments{
\item{example}{(character) One of \code{BMI} or \code{LDL} which have been included as example traits.}

\item{dataframe}{(logical) If \code{TRUE} (default), return an example munged dataframe. If \code{FALSE}, return path to the file on disk.}
}
\value{
either a \link[tibble:tibble-package]{tibble} containing a munged dataframe, or a path to the file on disk.
}
\description{
Example munged dataframe
}
\examples{
sumstats_munged_example(example = "BMI", dataframe = TRUE)
}
