% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.bicov.R
\name{ldsc.bicov}
\alias{ldsc.bicov}
\title{Single-Variate Linkage Disequilibrium Score Regression (LDSC)}
\usage{
ldsc.bicov(
  gwas1,
  gwas2,
  h1,
  h2,
  LDSC,
  zsquare_thresh = 50000,
  nblock = 500,
  sampling.time = 200,
  sampling.ratio = 0.5
)
}
\arguments{
\item{gwas1}{A data.frame containing GWAS summary statistics for the first trait. The data.frame should include columns for SNP identifiers, Z-scores of effect size estimates, sample sizes (N), effect allele (A1), and reference allele (A2).}

\item{gwas2}{A data.frame containing GWAS summary statistics for the second trait. The data.frame should include columns for SNP identifiers, Z-scores of effect size estimates, sample sizes (N), effect allele (A1), and reference allele (A2).}

\item{h1}{The heritability of the first trait.}

\item{h2}{The heritability of the second trait.}

\item{LDSC}{A data.frame containing LD Score Regression (LDSC) estimates. It should include LDSC scores and other necessary metrics for the analysis.}

\item{zsquare_thresh}{A threshold for the squared Z-scores in heritability estimation to control for extreme values.}

\item{nblock}{The number of blocks for bootstrap-based standard error estimation.}

\item{sampling.time}{The number of block bootstrap.}

\item{sampling.ratio}{The sub-sampling ratio in each bootstrap.}
}
\value{
A data.frame containing heritability estimates and their standard errors, along with the intercept and its standard error.
}
\description{
The `ldsc.bicov` function performs single-variate Linkage Disequilibrium Score Regression (LDSC) analysis. It is designed to estimate genetic covariance from two GWAS summary statistics, accounting for linkage disequilibrium (LD) between SNPs.
}
