% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_h2.R
\name{ldscore_files}
\alias{ldscore_files}
\title{Paths to LD-score files from Pan-UKB}
\usage{
ldscore_files(ancestry, ...)
}
\arguments{
\item{ancestry}{(character) One of "AFR", "AMR", "CSA", "EAS", "EUR", or "MID"}

\item{...}{arguments passed to \code{fs::dir_ls()}}
}
\value{
a list of paths
}
\description{
Paths to LD-score files from Pan-UKB
}
\examples{
ldscore_files("AFR")

}
