% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldscR-package.R
\docType{package}
\name{ldscR-package}
\alias{ldscR}
\alias{ldscR-package}
\title{ldscR: LD Score Regression in R}
\description{
This package implements LD score regression in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/ldscR}
  \item \url{https://mglev1n.github.io/ldscR/}
  \item Report bugs at \url{https://github.com/mglev1n/ldscR/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{mglev1n@gmail.com} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
