% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldscR.R
\name{ldscR}
\alias{ldscR}
\title{Estimate Heritability and Genetic Correlation Matrix Using LDSC estimated from sample LD matrix estimate}
\usage{
ldscR(GWAS_List, LDSC)
}
\arguments{
\item{GWAS_List}{A list of data.frames where each data.frame contains GWAS summary statistics for a trait. Each data.frame should include columns for SNP identifiers, Z-scores of effect size estimates, sample sizes (N), effect allele (A1), and reference allele (A2).}

\item{LDSC}{A data.frame containing LD Score Regression (LDSC) estimates. It should include LDSC scores and other necessary metrics for the analysis.}
}
\value{
A list containing the following elements:
\itemize{
  \item{GCovEst}{Estimated genetic covariance matrix.}
  \item{GCovSE}{Standard errors of the estimated genetic covariance matrix.}
  \item{ECovEst}{Estimated error covariance matrix.}
  \item{ECovSE}{Standard errors of the estimated error covariance matrix.}
}
}
\description{
The \code{ldscR} function estimates heritability and genetic correlation matrices using Linkage Disequilibrium Score Regression (LDSC). It processes GWAS summary statistics, harmonizes alleles with a reference panel, and computes genetic covariance and error covariance matrices.
}
\details{
The \code{ldscR} function is designed for advanced genetic statistics and requires a good understanding of GWAS summary statistics, LDSC methodology, and statistical genetics. Users should ensure that input data is correctly formatted and that they understand the implications of the estimates produced by the function.
}
\examples{
data(hapmap3)
data(EURLDSC)
ref_panel <- hapmap3
LDSC <- EURLDSC
results <- ldscR(GWAS_List, ref_panel, LDSC)

}
