% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldscR.R
\name{ldscR}
\alias{ldscR}
\title{Estimate Heritability and Genetic Correlation Matrix Using LDSC estimated from sample LD matrix estimate}
\usage{
ldscR(
  GWAS_List,
  LDSC,
  zsquare_thresh = 50000,
  cov_thresh = 10000,
  estimate_SE = F,
  sampling.time = 500,
  n_threads = NULL
)
}
\arguments{
\item{GWAS_List}{A list of data.frames where each data.frame contains GWAS summary statistics for a trait. Each data.frame should include columns for SNP identifiers, Z-scores of effect size estimates, sample sizes (N), effect allele (A1), and reference allele (A2).}

\item{LDSC}{A data.frame containing LD Score Regression (LDSC) estimates. It should include LDSC scores and other necessary metrics for the analysis.}

\item{zsquare_thresh}{Threshold on the z-score square in heritability estimation.}

\item{cov_thresh}{Threshold on the abs(z-score1 * z-score2) in genetic covariance estimation.}

\item{estimate_SE}{If estimating the standard error of the genetic covariance matrix and estimation error covariance matrix. Default to F.}

\item{sampling.time}{Bootstrapping time for the estimation of standard errors. Default to 200.}

\item{n_threads}{Number of threads used in Block-wise bootstrap. Default to NULL which uses half of CPU cores detected via OpenMP.}
}
\value{
A list containing the following elements:
\itemize{
  \item{GCovEst}{Estimated genetic covariance matrix.}
  \item{GCovSE}{Standard errors of the estimated genetic covariance matrix.}
  \item{ECovEst}{Estimated error covariance matrix.}
  \item{ECovSE}{Standard errors of the estimated error covariance matrix.}
  \item{Computing.time}{Computing time in each stage.}
}
}
\description{
The \code{ldscR} function estimates heritability and genetic correlation matrices using Linkage Disequilibrium Score Regression (LDSC). It processes GWAS summary statistics, harmonizes alleles with a reference panel, and computes genetic covariance and error covariance matrices.
}
